feat ajout create, delete, update service controller (sans log)

+++ src/services/services.controller.ts
@@ -1,15 +1,50 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  Res,
+  BadRequestException,
+  ParseIntPipe,
+  Bind,
+} from '@nestjs/common';
 import { ServicesService } from './services.service';
 import { CreateServiceDto } from './dto/create-service.dto';
 import { UpdateServiceDto } from './dto/update-service.dto';
+import { Response } from 'express';
+import { UsersService } from 'src/users/users.service';
 
 @Controller('services')
 export class ServicesController {
-  constructor(private readonly servicesService: ServicesService) {}
+  constructor(
+    private readonly servicesService: ServicesService,
+    private readonly usersService: UsersService,
+  ) {}
 
-  @Post()
-  create(@Body() createServiceDto: CreateServiceDto) {
-    return this.servicesService.create(createServiceDto);
+  @Post(':id')
+  @Bind(Param('id', new ParseIntPipe()))
+  async create(
+    id: string,
+    @Body() createServiceDto: CreateServiceDto,
+    @Res() res: Response,
+  ) {
+    //remplacer le parametre par l'id en parametre
+    const isUserExist = await this.usersService.findOne(+id);
+
+    if (!isUserExist) {
+      throw new BadRequestException('User id inconnu');
+    }
+
+    const newService = await this.servicesService.create(createServiceDto);
+
+    return res.status(201).json({
+      status: 'SUCCESS',
+      message: 'Nouveau service ajouté !',
+      data: newService,
+    });
   }
 
   @Get()
@@ -23,12 +58,41 @@ export class ServicesController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateServiceDto: UpdateServiceDto) {
-    return this.servicesService.update(+id, updateServiceDto);
+  @Bind(Param('id', new ParseIntPipe()))
+  async update(
+    @Param('id') id: string,
+    @Body() updateServiceDto: UpdateServiceDto,
+    @Res() res: Response,
+  ) {
+    const isUserExist = await this.usersService.findOne(+id);
+
+    if (!isUserExist) {
+      throw new BadRequestException('User id inconnu');
+    }
+    const updateService = await this.servicesService.update(
+      +id,
+      updateServiceDto,
+    );
+    return res.status(201).json({
+      status: 'SUCCESS',
+      message: 'Service modifié',
+      data: updateService,
+    });
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.servicesService.remove(+id);
+  @Bind(Param('id', new ParseIntPipe()))
+  async remove(@Param('id') id: string, @Res() res: Response) {
+    const isUserExist = await this.usersService.findOne(+id);
+
+    if (!isUserExist) {
+      throw new BadRequestException('User id inconnu');
+    }
+    const deleteService = await this.servicesService.remove(+id);
+    return res.status(201).json({
+      status: 'SUCCESS',
+      message: 'Service modifié',
+      data: deleteService,
+    });
   }
 }
